requirements.txt
requests
configparser
pyqt5
pyqt5-stubs
python-vlc
Pillow